monolith/src/main/webapp/app/entities/mst-district/mst-district.component.spec.ts
/* tslint:disable max-line-length */
import { vitest } from 'vitest';
import { shallowMount, type MountingOptions } from '@vue/test-utils';
import sinon, { type SinonStubbedInstance } from 'sinon';

import MstDistrict from './mst-district.vue';
import MstDistrictService from './mst-district.service';
import AlertService from '@/shared/alert/alert.service';

type MstDistrictComponentType = InstanceType<typeof MstDistrict>;

const bModalStub = {
  render: () => {},
  methods: {
    hide: () => {},
    show: () => {},
  },
};

describe('Component Tests', () => {
  let alertService: AlertService;

  describe('MstDistrict Management Component', () => {
    let mstDistrictServiceStub: SinonStubbedInstance<MstDistrictService>;
    let mountOptions: MountingOptions<MstDistrictComponentType>['global'];

    beforeEach(() => {
      mstDistrictServiceStub = sinon.createStubInstance<MstDistrictService>(MstDistrictService);
      mstDistrictServiceStub.retrieve.resolves({ headers: {} });

      alertService = new AlertService({
        i18n: { t: vitest.fn() } as any,
        bvToast: {
          toast: vitest.fn(),
        } as any,
      });

      mountOptions = {
        stubs: {
          jhiItemCount: true,
          bPagination: true,
          bModal: bModalStub as any,
          'font-awesome-icon': true,
          'b-badge': true,
          'jhi-sort-indicator': true,
          'b-button': true,
          'router-link': true,
        },
        directives: {
          'b-modal': {},
        },
        provide: {
          alertService,
          mstDistrictService: () => mstDistrictServiceStub,
        },
      };
    });

    describe('Mount', () => {
      it('Should call load all on init', async () => {
        // GIVEN
        mstDistrictServiceStub.retrieve.resolves({ headers: {}, data: [{ id: 123 }] });

        // WHEN
        const wrapper = shallowMount(MstDistrict, { global: mountOptions });
        const comp = wrapper.vm;
        await comp.$nextTick();

        // THEN
        expect(mstDistrictServiceStub.retrieve.calledOnce).toBeTruthy();
        expect(comp.mstDistricts[0]).toEqual(expect.objectContaining({ id: 123 }));
      });

      it('should calculate the sort attribute for an id', async () => {
        // WHEN
        const wrapper = shallowMount(MstDistrict, { global: mountOptions });
        const comp = wrapper.vm;
        await comp.$nextTick();

        // THEN
        expect(mstDistrictServiceStub.retrieve.lastCall.firstArg).toMatchObject({
          sort: ['id,asc'],
        });
      });
    });
    describe('Handles', () => {
      let comp: MstDistrictComponentType;

      beforeEach(async () => {
        const wrapper = shallowMount(MstDistrict, { global: mountOptions });
        comp = wrapper.vm;
        await comp.$nextTick();
        mstDistrictServiceStub.retrieve.reset();
        mstDistrictServiceStub.retrieve.resolves({ headers: {}, data: [] });
      });

      it('should load a page', async () => {
        // GIVEN
        mstDistrictServiceStub.retrieve.resolves({ headers: {}, data: [{ id: 123 }] });

        // WHEN
        comp.page = 2;
        await comp.$nextTick();

        // THEN
        expect(mstDistrictServiceStub.retrieve.called).toBeTruthy();
        expect(comp.mstDistricts[0]).toEqual(expect.objectContaining({ id: 123 }));
      });

      it('should not load a page if the page is the same as the previous page', () => {
        // WHEN
        comp.page = 1;

        // THEN
        expect(mstDistrictServiceStub.retrieve.called).toBeFalsy();
      });

      it('should re-initialize the page', async () => {
        // GIVEN
        comp.page = 2;
        await comp.$nextTick();
        mstDistrictServiceStub.retrieve.reset();
        mstDistrictServiceStub.retrieve.resolves({ headers: {}, data: [{ id: 123 }] });

        // WHEN
        comp.clear();
        await comp.$nextTick();

        // THEN
        expect(comp.page).toEqual(1);
        expect(mstDistrictServiceStub.retrieve.callCount).toEqual(1);
        expect(comp.mstDistricts[0]).toEqual(expect.objectContaining({ id: 123 }));
      });

      it('should calculate the sort attribute for a non-id attribute', async () => {
        // WHEN
        comp.propOrder = 'name';
        await comp.$nextTick();

        // THEN
        expect(mstDistrictServiceStub.retrieve.lastCall.firstArg).toMatchObject({
          sort: ['name,asc', 'id'],
        });
      });

      it('Should call delete service on confirmDelete', async () => {
        // GIVEN
        mstDistrictServiceStub.delete.resolves({});

        // WHEN
        comp.prepareRemove({ id: 123 });

        comp.removeMstDistrict();
        await comp.$nextTick(); // clear components

        // THEN
        expect(mstDistrictServiceStub.delete.called).toBeTruthy();

        // THEN
        await comp.$nextTick(); // handle component clear watch
        expect(mstDistrictServiceStub.retrieve.callCount).toEqual(1);
      });
    });
  });
});
